Reviewing tests
- Rename manage, action as per the rest of my Apps
- Reviewed resource type tests
